Create context to manage server state
